Add-on: Add checkbox to submit jobs in `paused` status (#104323)
Add a checkbox to the submission interface to submit the job in
`paused` status.

Internally this translates the boolean value of the checkbox to the
desired state (`queued` or `paused`). A checkbox was chosen to
simplify the user interface; technically a drop-down could have been
used to select the initial state, but this would require more clicks
to get the same result.

Pull Request: https://projects.blender.org/studio/flamenco/pulls/104323