Worker: improve warning when FFmpeg cannot be found
On Linux (at least my Ubuntu) the error from finding FFmpeg wasn't
identified as a "file not found" error, and thus the more abstract
"unexpected error" message was logged.

Now any error with "file not found" in the message is handled as such,
and just produces a warning without the word "error" in there.