Subdivide tasks into steps, and let the worker publish its step progress
Tasks and Jobs now keep track of how many "steps" they contain in total,
and how many have been completed so far. This is shown in the web
frontend as a little progress bar underneath the task/job status.

At the lowest level, it is the Command authoring that declares how many
steps that Command should take. The number of steps for each Task is the
sum of its Commands; and similarly the number of steps in a Job is the
sum of its Tasks.

The Command implementation on the Worker has to support steps. Currently
the only command that supports this is `blender-render`; it takes each
rendered frame as a step. It is up to the job compiler JavaScript code
to tell the command how many frames are expected to render (this can be
tricky to predict automatically given that arbitrary Python code can be
passed to the command).

Each Command that was not authored with an explicit number of steps is
counted as a single step. This includes commands that do support steps,
but for which the job compiler has not given any step count. This is for
backward compatibility, making existing JS files work like they did
before. The screenshot shows a task of 20 `sleep` commands, and thus
that task consists of 20 steps.

In Flamenco's current design, it is just the JS script and the Worker
that need knowledge of specific commands. The Manager is just moving
info back & forth without knowing what these commands mean. I like that
design. Of course it would be possible to make the Manager more aware of
which command takes how many steps, given its parameters. However, that
would necessarily give the Manager knowledge about the commands. This in
turn would make changes to commands more cumbersome and error-prone, as
it would need updates to three areas (JS, Manager, and Worker). Limiting
the changes to the parts that author (JS) or execute (Worker) the
commands seems more sensible to me.