WebApp: Cast FormInputNumber value to Number onInput
Fix #104432, where the form throws an error and cannot be saved. It is
happening because the backend is expecting `number`, not `string`, and
whenever the user types, the value in the input gets changed to a
string. The solution is to cast `FormInputNumber` value to a number
`onInput` to always maintain the `number` type.

Pull Request: https://projects.blender.org/studio/flamenco/pulls/104434