Webapp: don't fetch entire job on receiving a job update via socketIO
When a SocketIO job update comes in, just update the jobs table. This
reduces the number of updates considerably, because now each task step
also updates its job step count, and so job updates are much more
frequent.

In the future the job update itself could indicate whether it was for
changes outside the job summary, and thus re-fetching the job is still
necessary.