Webapp: Show banner on settings form submission
Show a banner on submission of the settings form.

On success, the banner displays `Settings successfully saved.` and
disappears after 3s or if the user clicks the X.

On fail, the banner displays `Failed to save: ${errorMessage}` where
the error message is parsed from `error.body.message` from the catch
body of `this.metaAPI.updateConfigurationFile()`. It remains on the
screen until manually closed by the user by clicking the X.

Pull Request: https://projects.blender.org/studio/flamenco/pulls/104441