Manager: always run queries in an explicit RO or RW transaction
Instead of using the SQLite auto-commit feature, always explicitly start
a transaction using a function that's suitable for read-only queries or
read-write queries.

This also introduces a read-write mutex on the Go side of things
(instead of relying on SQLite's locking/conflict management system).
This makes it significantly easier to ensure there's only one writing
transaction at a time.

The mutex makes it possible to increase the number of database
connections, as now multiple read-only connections can be allowed in
parallel. Because the connections are now made by the `database/sql`
package, the `PRAGMA` queries that were previously passed explicitly are
now passed via the database URL. That way each new connection gets the
same configuration.