Fix #104459: Saving settings without restarting corrupts in-memory config
After saving the new settings, process the settings as if they were just
loaded from disk. This ensures that internal, runtime-only fields are
also set correctly.

To be able to test this properly, this commit also extends the config
service with control (for tests) over the filename the config is loaded
from / saved to.