Upgrade Go tools & modernize how they are invoked
- Move these tools from the quirky `tools.go` we had to proper `go tool`
  use:
    - mockgen
    - mage
    - oapi-codegen
- Upgrade oapi-codegen v1.15.0 → 2.5.1

See [1] for more on `go tool`.

The update to oapi-codegen is quite invasive, as it's a major update and
there have been some changes to the generated code. IMO it's all for the
better, so it's a welcome improvement. The upgrade was also a necesity
to get the package to work with the `go tool` approach.

The other tools have not been upgraded, it's just that we install &
invoke them differently.

[1] https://www.jvt.me/posts/2025/01/27/go-tools-124/