Manager: speed up startup by being smarter about database migrations
Investigate the database schema version, and if there are no pending
migrations to perform, skip the following steps:

- Disable foreign key constraints
- Perform the migration
- Enable foreign key constraints
- Perform integrity check
- Vacuum the database

This halves the startup time of Flamenco Manager when the database gets
larger. At startup there is still always an integrity check + vacuum
pass (it's the second post-migration ones that are skipped now).