Cleanup: explicit error checking / ignoring
Add explicit error checks (mostly on deferred 'close' calls) or explicit
error ignores.

Also I replaced file more complex file writing with writing to an
in-memory buffer (which always succeeds) and a single-call writing of
that buffer to file (which is checked).

Note: this exposes an actual bug, which will be fixed in the next
commit.

# Conflicts:
#	internal/manager/task_logs/task_logs.go